+++ test/jdk/java/lang/invoke/TestFindClass.java
@@ -26,9 +26,9 @@
 /* @test
  * @bug 8150782 8207027 8266269
  * @compile TestFindClass.java TestCls.java p/Foo.java q/Bar.java
- * @run testng/othervm -ea -esa test.java.lang.invoke.t8150782.TestFindClass
+ * @run testng/othervm -ea -esa test.java.lang.invoke.TestFindClass
  */
-package test.java.lang.invoke.t8150782;
+package test.java.lang.invoke;
 
 import java.lang.invoke.*;
 import p.Foo;
@@ -42,7 +42,7 @@ import org.testng.annotations.*;
 
 public class TestFindClass {
 
-    private static final String PACKAGE_PREFIX = "test.java.lang.invoke.t8150782.";
+    private static final String PACKAGE_PREFIX = "test.java.lang.invoke.";
 
     private static boolean initializedClass1;
