+++ test/jdk/java/lang/invoke/p/Foo.java
