+++ test/jdk/java/lang/invoke/q/Bar.java
